Move all host selectors to an own file
