+++ assets/js/components/InstanceOverview/InstanceOverview.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { getHost } from '@state/selectors';
+import { getHost } from '@state/selectors/host';
 import { getCluster } from '@state/selectors/cluster';
 import HealthIcon from '@components/Health';
 import { Features } from '@components/SapSystemDetails';
